Move playback code to a separate file. It is about to become a bit more complicated due to rewriting linked object BLOBs.
