+++ services/src/mrnet/MRNet_Send.c
@@ -27,19 +27,9 @@
 #include "config.h"
 #endif
 
-#include <errno.h>
 #include <pthread.h>
-#include <stdbool.h>
-#include <stdint.h>
-#include <stdio.h>
-#include <stdlib.h>
-#include <sys/types.h>
-#include <sys/stat.h>
-#include <unistd.h>
 
-#include "KrellInstitute/Services/Assert.h"
 #include "KrellInstitute/Services/Common.h"
-#include "KrellInstitute/Services/Time.h"
 #include "KrellInstitute/Messages/DataHeader.h"
 #include "KrellInstitute/Messages/EventHeader.h"
 #include "KrellInstitute/Messages/Blob.h"
@@ -50,6 +40,10 @@
 
 #include "KrellInstitute/CBTF/Impl/MessageTags.h"
 
+#if defined(ENABLE_CBTF_MRNET_PLAYBACK)
+#include "playback.h"
+#endif
+
 Network_t* CBTF_MRNet_netPtr;
 // make the id of the stream we want global and use Network_get_Stream
 // locally in the send function to retrieve it.
@@ -59,158 +53,6 @@ static int mrnet_connected = 0;
 
 static pthread_mutex_t mrnet_connected_mutex = PTHREAD_MUTEX_INITIALIZER;
 static pthread_cond_t mrnet_connected_cond = PTHREAD_COND_INITIALIZER;
-
-static bool do_record_playback = false;
-static bool do_replay_playback = false;
-static bool no_playback_delay = false;
-static bool is_first_message = true;
-static bool do_squelch_message = false;
-
-static char playback_file[PATH_MAX] = "";
-
-
-
-static void CBTF_MRNet_LW_sendToFrontend(const int tag,
-                                         const int size, void *data);
-
-
-
-static void configure_playback()
-{
-    do_record_playback = (getenv("CBTF_MRNET_RECORD_PLAYBACK") != NULL);
-    do_replay_playback = (getenv("CBTF_MRNET_REPLAY_PLAYBACK") != NULL);
-    no_playback_delay = (getenv("CBTF_MRNET_NO_PLAYBACK_DELAY") != NULL);
-    
-    if (do_record_playback && do_replay_playback)
-    {
-        do_record_playback = false;
-        do_replay_playback = false;
-
-        fprintf(stderr, "CBTF_MRNet_LW_connect: "
-                "cannot both record and replay a playback concurrently!\n");
-        fflush(stderr);
-    }
-    else if (do_record_playback || do_replay_playback)
-    {
-        char* directory = do_record_playback ?
-            getenv("CBTF_MRNET_RECORD_PLAYBACK") :
-            getenv("CBTF_MRNET_REPLAY_PLAYBACK");
-        
-        struct stat status;
-        int retval = stat(directory, &status);
-
-        if ((retval == -1) || (!S_ISDIR(status.st_mode)))
-        {
-            do_record_playback = false;
-            do_replay_playback = false;
-
-            fprintf(stderr, "CBTF_MRNet_LW_connect: "
-                    "playback directory \"%s\" couldn't be accessed!\n",
-                    directory);
-            fflush(stderr);
-        }
-        else
-        {
-            sprintf(playback_file, "%s/cbtf-mrnet-playback-%u",
-                    directory, Network_get_LocalRank(CBTF_MRNet_netPtr));
-
-            retval = stat(playback_file, &status);
-
-            if (do_record_playback &&
-                ((retval == 0) || (errno != ENOENT)))
-            {
-                do_record_playback = false;
-                
-                fprintf(stderr, "CBTF_MRNet_LW_connect: "
-                        "playback file \"%s\" already exists!\n",
-                        playback_file);
-                fflush(stderr);
-            }
-            
-            if (do_replay_playback &&
-                ((retval == -1) || !S_ISREG(status.st_mode)))
-            {
-                do_replay_playback = false;
-                
-                fprintf(stderr, "CBTF_MRNet_LW_connect: "
-                        "playback file \"%s\" couldn't be opened!\n",
-                        playback_file);
-                fflush(stderr);
-            }
-        }
-    }
-}
-
-
-
-static void record_to_playback(int32_t tag, uint32_t size, void* data)
-{
-    uint64_t time = CBTF_GetTime();
-    FILE* fp = fopen(playback_file, "a+");
-
-    Assert(fp != NULL);
-
-    Assert(fwrite(&time, sizeof(uint64_t), 1, fp) == 1);
-    Assert(fwrite(&tag, sizeof(int32_t), 1, fp) == 1);
-    Assert(fwrite(&size, sizeof(uint32_t), 1, fp) == 1);
-    Assert(fwrite(data, size, 1, fp) == 1);
-    
-    Assert(fclose(fp) == 0);
-}
-
-
-
-static void replay_from_playback()
-{
-    bool is_first = true;
-    uint64_t previous_time = 0;
-    FILE* fp = fopen(playback_file, "r");
-
-    Assert(fp != NULL);
-
-    while (true)
-    {
-        uint64_t time = 0;
-        int32_t tag = 0;
-        uint32_t size = 0;
-
-        if (fread(&time, sizeof(uint64_t), 1, fp) < 1)
-        {
-            break;
-        }
-
-        if (fread(&tag, sizeof(int32_t), 1, fp) < 1)
-        {
-            break;
-        }
-
-        if (fread(&size, sizeof(uint32_t), 1, fp) < 1)
-        {
-            break;
-        }
-
-        void* data = malloc(size);
-
-        if (fread(data, size, 1, fp) < 1)
-        {
-            free(data);
-            break;
-        }
-
-        if (!is_first && !no_playback_delay)
-        {
-            usleep((time - previous_time) / 1000 /* us/ns */);
-        }
-
-        is_first = false;
-        previous_time = time;
-
-        CBTF_MRNet_LW_sendToFrontend(tag, (int)size, data);
-        free(data);
-    }
-        
-    Assert(fclose(fp) == 0);
-}
 
 
 
@@ -267,6 +109,8 @@ static int CBTF_MRNet_getParentInfo(const char* file, int rank, char* phost, cha
 
     return 1;
 }
+
+
 
 int CBTF_MRNet_LW_connect (const int con_rank)
 {
@@ -406,13 +250,10 @@ int CBTF_MRNet_LW_connect (const int con_rank)
 #endif
      }
 
-    /*
-     * Determine if the sent messages should be recorded for future playback,
-     * or replayed from an existing playback. And, if so, check for the given
-     * playback directory and determine the name of the proper playback file.
-     */
-    configure_playback();
-
+#if defined(ENABLE_CBTF_MRNET_PLAYBACK)
+    configure_playback(Network_get_LocalRank(CBTF_MRNet_netPtr));
+#endif
+    
     mrnet_connected = 1;
     pthread_cond_broadcast(&mrnet_connected_cond);
     pthread_mutex_unlock(&mrnet_connected_mutex);
@@ -423,7 +264,9 @@ int CBTF_MRNet_LW_connect (const int con_rank)
     return 1;
 }
 
-static void CBTF_MRNet_LW_sendToFrontend(const int tag, const int size, void *data)
+
+
+void CBTF_MRNet_LW_sendToFrontend(const int tag, const int size, void *data)
 {
     const char* fmt_str = "%auc";
 
@@ -439,33 +282,13 @@ static void CBTF_MRNet_LW_sendToFrontend(const int tag, const int size, void *da
     }
     pthread_mutex_unlock(&mrnet_connected_mutex);
 
-    /* Record this message to a playback file if requested. */
-    if (do_record_playback)
+#if defined(ENABLE_CBTF_MRNET_REPLAY)
+    if (playback_intercept(tag, (uint32_t)size, data))
     {
-        record_to_playback(tag, (uint32_t)size, data);
+        return; /* Squelch the actual sending of this message. */
     }
-
-    /*
-     * Squelch the actual sending of this message if replay of a playback file
-     * has been requested. But if this is the first message, do the replay too.
-     * Things get slightly complicated because replay_from_playback() must use
-     * this same function to send its messages...     
-     */
-    if (do_replay_playback)
-    {
-        if (is_first_message)
-        {
-            is_first_message = false;
-            replay_from_playback();
-            do_squelch_message = true;
-        }
-
-        if (do_squelch_message)
-        {
-            return;
-        }
-    }
-
+#endif
+    
 #ifndef NDEBUG
     if (getenv("CBTF_DEBUG_LW_MRNET") != NULL) {
 	fprintf(stderr,"CBTF_MRNet_LW_sendToFrontend: sends message with tag %d\n",tag);
@@ -483,8 +306,9 @@ static void CBTF_MRNet_LW_sendToFrontend(const int tag, const int size, void *da
     fflush(stderr);
 }
 
-void CBTF_MRNet_Send(const int tag,
-                 const xdrproc_t xdrproc, const void* data)
+
+
+void CBTF_MRNet_Send(const int tag, const xdrproc_t xdrproc, const void* data)
 {
     unsigned size=0,dm_size=0;
     char* dm_contents = NULL;
@@ -511,8 +335,10 @@ void CBTF_MRNet_Send(const int tag,
     CBTF_MRNet_LW_sendToFrontend(tag ,dm_size , (void *) dm_contents);
 }
 
+
+
 void CBTF_MRNet_Send_PerfData(const CBTF_DataHeader* header,
-                 const xdrproc_t xdrproc, const void* data)
+                              const xdrproc_t xdrproc, const void* data)
 {
     const size_t EncodingBufferSize = (CBTF_BlobSizeFactor * 15 * 1024);
     unsigned size;
@@ -541,7 +367,10 @@ void CBTF_MRNet_Send_PerfData(const CBTF_DataHeader* header,
                  (xdrproc_t) xdr_CBTF_Protocol_Blob, &blob);
 }
 
-void CBTF_Waitfor_MRNet_Shutdown() {
+
+
+void CBTF_Waitfor_MRNet_Shutdown()
+{
 
     Packet_t * p;
     p = (Packet_t *)malloc(sizeof(Packet_t));
